refactor(checklists): move the three pure utils into the engine (v3 Phase 2, 2.1)
`dependencies.ts`, `progress.ts` and `create-checklist.ts` move to
`lib/checklists-engine/`, re-typed against the engine's types instead of the
React barrel. `src/utils/*.ts` become one-line re-exports, so `src/index.ts`
and the three util specs never move.

Re-typing is generic rather than widening, so the React types still narrow
through: `resolveTaskDependencies<TTask>`, `getNextTask<TConfig>` and
`getLockedTasks<TConfig>` return the caller's own task type, and
`createChecklist<TConfig>`/`createTask<TTask>` return their argument's type
rather than the base config.

`logger` now comes from `@tour-kit/core/engine`, not bare `@tour-kit/core`:
the bare barrel is React-bearing and would land in the /engine closure.

Recipe row 14 (the duplicate an extraction exposes) checked and clean — all
eight moved functions have exactly one definition in the package. Unlike
`hints`, where `getHotspotPosition` was shipped twice.

Note for the outcome: the plan's Task 2.1 sanity grep over `src/lib` cannot
pass as written. `src/lib/slot.tsx` is a pre-existing UnifiedSlot barrel
(PR #10) that imports bare core by design. The guard that matters is scoped to
`src/lib/checklists-engine/`, which is clean.

24 files / 339 tests green; typecheck clean.
